requirements.txt
Django==4.0.3
djangorestframework==3.13.1
psycopg2-binary==2.9.3
Pillow==9.1.0